Add files via upload
Check the validity of the New Keynesian Phillips Curve in the Indian Context. Reverse engineer the values of real marginal cost for the 1960-2000. On the basis of the calculated values for the train data, predict values for Valid. If it fits well, continue prediction for forecast years. Then calculate inflation using NKPC. Check the values of actual inflation against forecasted inflation to examine the validity of the model.